requirements.txt
apache-airflow==2.10.5
duckdb==1.2.2
psycopg2-binary